travelLog BudgetView renamed, travelLogView create, travelLogPostIt Cell create
