+++ OGADA/OGADA/View/TravelLog/PostItCollectionViewCell.swift
@@ -1,0 +1,23 @@
+//
+//  TravelLogPostItCollectionViewCell.swift
+//  OGADA
+//
+//  Created by YoujinMac on 2020/03/09.
+//  Copyright © 2020 didwndckd. All rights reserved.
+//
+
+import UIKit
+
+class PostItCollectionViewCell: UICollectionViewCell {
+    
+//    private let day: Int
+//    private let date: String
+//    private let foreignCash: Int
+//    private let krwCash: Int
+//    private let foreignCard: Int
+//    private let krwCard: Int
+//    private let foreignTotal: Int
+//    private let krwTotal: Int
+    
+    
+}
