--- OGADA/OGADA/View/TravelLog/TravelLogPostItCollectionViewCell.swift
@@ -1,16 +1,0 @@
-//
-//  TravelLogPostItCollectionViewCell.swift
-//  OGADA
-//
-//  Created by YoujinMac on 2020/03/09.
-//  Copyright © 2020 didwndckd. All rights reserved.
-//
-
-import UIKit
-
-class TravelLogPostItCollectionViewCell: UICollectionViewCell {
-    
-    
-    
-    
-}
